// src/app/ingestors/discord/discord-api-types.ts
export interface components {
    schemas: {
// ...
        MessageAttachmentResponse: {
            id: components["schemas"]["SnowflakeType"];
            filename: string;
            /** Format: int32 */
            size: number;
            /** Format: uri */
            url: string;
            /** Format: uri */
            proxy_url: string;
            /** Format: int32 */
            width?: number;
            /** Format: int32 */
            height?: number;
            /** Format: double */
            duration_secs?: number;
            waveform?: string;
            description?: string;
            content_type?: string;
            ephemeral?: boolean;
            title?: string | null;
            application?: components["schemas"]["ApplicationResponse"];
            /** Format: date-time */
            clip_created_at?: string;
            clip_participants?: components["schemas"]["UserResponse"][];
        };
// ...
        MessageResponse: {
            type: components["schemas"]["MessageType"];
            content: string;
            mentions: components["schemas"]["UserResponse"][];
            mention_roles: components["schemas"]["SnowflakeType"][];
            attachments: components["schemas"]["MessageAttachmentResponse"][];
            embeds: components["schemas"]["MessageEmbedResponse"][];
            /** Format: date-time */
            timestamp: string;
            /** Format: date-time */
            edited_timestamp?: string | null;
            /** Format: int32 */
            flags: number;
            components: (components["schemas"]["ActionRowComponentResponse"] | components["schemas"]["ContainerComponentResponse"] | components["schemas"]["FileComponentResponse"] | components["schemas"]["MediaGalleryComponentResponse"] | components["schemas"]["SectionComponentResponse"] | components["schemas"]["SeparatorComponentResponse"] | components["schemas"]["TextDisplayComponentResponse"])[];
            stickers?: (components["schemas"]["GuildStickerResponse"] | components["schemas"]["StandardStickerResponse"])[];
            sticker_items?: components["schemas"]["MessageStickerItemResponse"][];
            id: components["schemas"]["SnowflakeType"];
            channel_id: components["schemas"]["SnowflakeType"];
            author: components["schemas"]["UserResponse"];
            pinned: boolean;
            mention_everyone: boolean;
            tts: boolean;
            call?: components["schemas"]["MessageCallResponse"];
            activity?: components["schemas"]["MessageActivityResponse"];
            application?: components["schemas"]["BasicApplicationResponse"];
            application_id?: components["schemas"]["SnowflakeType"];
            interaction?: components["schemas"]["MessageInteractionResponse"];
            nonce?: number | string | null;
            webhook_id?: components["schemas"]["SnowflakeType"];
            message_reference?: components["schemas"]["MessageReferenceResponse"];
            thread?: components["schemas"]["ThreadResponse"];
            mention_channels?: (null | components["schemas"]["MessageMentionChannelResponse"])[];
            role_subscription_data?: components["schemas"]["MessageRoleSubscriptionDataResponse"];
            purchase_notification?: components["schemas"]["PurchaseNotificationResponse"];
            /** Format: int32 */
            position?: number;
            resolved?: components["schemas"]["ResolvedObjectsResponse"];
            poll?: components["schemas"]["PollResponse"];
            shared_client_theme?: null | components["schemas"]["CustomClientThemeResponse"];
            interaction_metadata?: components["schemas"]["ApplicationCommandInteractionMetadataResponse"] | components["schemas"]["MessageComponentInteractionMetadataResponse"] | components["schemas"]["ModalSubmitInteractionMetadataResponse"];
            message_snapshots?: components["schemas"]["MessageSnapshotResponse"][];
            reactions?: components["schemas"]["MessageReactionResponse"][];
            referenced_message?: null | components["schemas"]["BasicMessageResponse"];
        };
// ...
        UserResponse: {
            id: components["schemas"]["SnowflakeType"];
            username: string;
            avatar?: string | null;
            discriminator: string;
            /** Format: int32 */
            public_flags: number;
            flags: components["schemas"]["Int53Type"];
            bot?: boolean;
            system?: boolean;
            banner?: string | null;
            /** Format: int32 */
            accent_color?: number | null;
            global_name?: string | null;
            avatar_decoration_data?: null | components["schemas"]["UserAvatarDecorationResponse"];
            collectibles?: null | components["schemas"]["UserCollectiblesResponse"];
            primary_guild?: null | components["schemas"]["UserPrimaryGuildResponse"];
        };
// ...
    };
// ...
}